* commit '0a2bb11b248df6841daabca3970df5d8504adfc7':
  address problems from comments
  fix and address comments
  change according to previous comments
  add missing piece
  remove unnecessary check macro
  fix intel compile warnings
  Revert "fix warnings from Intel compiler"
  Revert "fix warnings and some text alignment"
  Revert "let hdf5 pick up the right compiler in Intel environment"
  Revert "fix issues from previous PR comments"
  Revert "using a different MACRO"
  using a different MACRO
  fix issues from previous PR comments
  let hdf5 pick up the right compiler in Intel environment
  fix warnings and some text alignment
  fix warnings from Intel compiler

+++ tools/h5dump/dynlib_dump.c
@@ -69,7 +69,7 @@ H5Z_filter_dynlibud(unsigned int flags, size_t cd_nelmts,
         /* Subtract the original value with MULTIPLIER */
         while(buf_left > 0) {
             char temp = *int_ptr;
-            *int_ptr = temp - MULTIPLIER;
+            *int_ptr = (int8_t)(temp - MULTIPLIER);
             int_ptr++;
             buf_left -= sizeof(*int_ptr);
         } /* end while */
@@ -78,7 +78,7 @@ H5Z_filter_dynlibud(unsigned int flags, size_t cd_nelmts,
         /* Add the original value with MULTIPLIER */
         while(buf_left > 0) {
             char temp = *int_ptr;
-            *int_ptr = temp + MULTIPLIER;
+            *int_ptr = (int8_t)(temp + MULTIPLIER);
             int_ptr++;
             buf_left -= sizeof(*int_ptr);
         } /* end while */
